# karma ai/api/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "BuildAcceptedDTO": {
        "properties": {
          "build_id": {
            "title": "Build Id",
            "type": "string"
          },
          "poll_after_ms": {
            "default": 2000,
            "title": "Poll After Ms",
            "type": "integer"
          },
          "status": {
            "const": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "build_id",
          "status"
        ],
        "title": "BuildAcceptedDTO",
        "type": "object"
      },
      "BuildCardDTO": {
        "properties": {
          "parts": {
            "items": {
              "$ref": "#/components/schemas/BuildPartDTO"
            },
            "title": "Parts",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "total_price_inr": {
            "title": "Total Price Inr",
            "type": "integer"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "parts",
          "total_price_inr",
          "summary"
        ],
        "title": "BuildCardDTO",
        "type": "object"
      },
      "BuildPartDTO": {
        "properties": {
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "justification": {
            "title": "Justification",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_inr": {
            "title": "Price Inr",
            "type": "integer"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "slot": {
            "title": "Slot",
            "type": "string"
          }
        },
        "required": [
          "slot",
          "product_id",
          "name",
          "price_inr",
          "justification"
        ],
        "title": "BuildPartDTO",
        "type": "object"
      },
      "BuildStatusResponse": {
        "properties": {
          "build": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BuildCardDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "build_id": {
            "title": "Build Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "poll_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poll After Ms"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "infeasible",
              "cannot_proceed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "build_id",
          "status"
        ],
        "title": "BuildStatusResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartBuildRequest": {
        "properties": {
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "StartBuildRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerdictDTO": {
        "properties": {
          "binding_constraint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Binding Constraint"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "suggested_adjustments": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Suggested Adjustments",
            "type": "array"
          },
          "verdict": {
            "enum": [
              "comfortable",
              "tight",
              "impossible"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "verdict",
          "reason"
        ],
        "title": "VerdictDTO",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Karma Computers"
    },
    "description": "HTTP layer over the Karma Advisor LangGraph pipeline — conversational intake, feasibility, budget allocation, and part selection. The frozen frontend contract (DTOs, error catalog, per-screen fixtures) is documented in docs/frontend_contract_plan.md.",
    "title": "Karma Advisor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/builds": {
      "post": {
        "operationId": "start_build_api_v1_builds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildAcceptedDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHORIZED — missing or invalid X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "SESSION_NOT_FOUND — unknown or expired session."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "BRIEF_NOT_LOCKED — the session's brief is not locked yet."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "VALIDATION_ERROR — request body failed validation."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "BUILD_CAPACITY (at max concurrent builds) or RATE_LIMITED (build-create quota exceeded); both set Retry-After."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Build",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}": {
      "get": {
        "operationId": "get_build_api_v1_builds__build_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHORIZED — missing or invalid X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "BUILD_NOT_FOUND — unknown or evicted build id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Build",
        "tags": [
          "builds"
        ]
      }
    },
...
  }
}